Add unwarpKey to typescript interface
